Close #1 - Rewrite file parser to read infos from mod.info file
Removed latest_version.txt and tiny_avc.txt and added the necessary tags
to the mod.info file. Moved tag and delimiter definitions to constants
outside of the function.

Updated the downloadUpdates() function to first load the tags from the
local mod.info file. It will then use the modurl tag in that file to
grab a file online which contains the version infos.